server/requirements.txt
flask
zappa
flask_cors
mysql.connector
credentials